chore: moved highlight logic into standalone hook
