+++ src/components/FeatureInformation/FeatureInformation.js
@@ -1,28 +1,19 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 import { MdClose } from 'react-icons/md';
-import GeometryType from 'ol/geom/GeometryType';
 import VectorLayer from 'ol/layer/Vector';
 import VectorSource from 'ol/source/Vector';
-import Feature from 'ol/Feature';
-import Point from 'ol/geom/Point';
 import { IoIosArrowRoundBack, IoIosArrowRoundForward } from 'react-icons/io';
 import { Link, IconButton } from '@material-ui/core';
 import { setFeatureInfo } from '../../model/app/actions';
-import popups from '../../popups';
 import highlightPointStyle from '../../utils/highlightPointStyle';
-import panCenterFeature from '../../utils/panCenterFeature';
-
+import useIndexedFeatureInfo from '../../utils/useIndexedFeatureInfo';
+import useHighlightLayer from '../../utils/useHighlightLayer';
+import getPopupComponent from '../../utils/getPopupComponent';
 import './FeatureInformation.scss';
-import useIsMobile from '../../utils/useIsMobile';
-
-const getPopupComponent = ({ popupComponent, layer }) => {
-  const comp = popupComponent || layer.get('popupComponent');
-  return typeof comp === 'string' ? popups[comp] : comp;
-};
 
 const highlightLayer = new VectorLayer({
   source: new VectorSource({ features: [] }),
@@ -32,170 +23,16 @@ highlightLayer.setStyle(highlightPointStyle);
 const FeatureInformation = ({ featureInfo }) => {
   const { t } = useTranslation();
   const dispatch = useDispatch();
-  const map = useSelector((state) => state.app.map);
   const language = useSelector((state) => state.app.language);
-  const searchService = useSelector((state) => state.app.searchService);
-  const menuOpen = useSelector((state) => state.app.menuOpen);
   const appBaseUrl = useSelector((state) => state.app.appBaseUrl);
   const staticFilesUrl = useSelector((state) => state.app.staticFilesUrl);
   const [featureIndex, setFeatureIndex] = useState(0);
-  const isMobile = useIsMobile();
-
-  useEffect(() => {
-    // The featureInformation component can be display twice at the same time (in the popup and in the overlay).
-    // So to avoid a js error we have to check if the layer is already on the map or not.
-    if (
-      map &&
-      !map
-        .getLayers()
-        .getArray()
-        .find((layer) => layer === highlightLayer)
-    ) {
-      map.addLayer(highlightLayer);
-      // Clear the highlight for search
-      if (searchService) {
-        searchService.clearHighlight();
-      }
-    }
-
-    return () => {
-      highlightLayer.getSource().clear();
-
-      if (
-        map &&
-        map
-          .getLayers()
-          .getArray()
-          .find((layer) => layer === highlightLayer)
-      ) {
-        map.removeLayer(highlightLayer);
-      }
-    };
-  }, [map, searchService]);
-
-  useEffect(() => {
-    setFeatureIndex(0);
-  }, [featureInfo]);
 
   // List of features and layers available for pagination.
-  const infoIndexed = useMemo(() => {
-    const features = [];
+  const infoIndexed = useIndexedFeatureInfo(featureInfo);
 
-    // List of corresponding layer for each features in the array.
-    const layers = [];
-
-    // List of corresponding coordinates clicked for each features in the array.
-    const coordinates = [];
-
-    // When a popup use hidePagination, we store the index for each popup.
-    const indexByPopup = {};
-
-    featureInfo.forEach((featInfo) => {
-      const PopupComponent = getPopupComponent(featInfo);
-
-      if (PopupComponent && PopupComponent.hidePagination) {
-        const name = PopupComponent.displayName;
-        // All features using this PopupComponent will be render on the same page
-        if (indexByPopup[name] !== undefined) {
-          features[indexByPopup[name]].push(...featInfo.features);
-          featInfo.features.forEach(() => {
-            if (!layers[indexByPopup[name]]) {
-              layers[indexByPopup[name]] = [];
-              coordinates[indexByPopup[name]] = [];
-            }
-            if (!coordinates[indexByPopup[name]]) {
-              coordinates[indexByPopup[name]] = [];
-            }
-            layers[indexByPopup[name]].push(featInfo.layer);
-            coordinates[indexByPopup[name]].push(featInfo.coordinate);
-          });
-        } else {
-          // At this point features must be displayed in the same popup, that's why we push an array.
-          features.push([...featInfo.features]);
-          const arr = [];
-          const arrCoord = [];
-          featInfo.features.forEach(() => {
-            arr.push(featInfo.layer);
-            arrCoord.push(featInfo.coordinate);
-          });
-          layers.push(arr);
-          coordinates.push(arrCoord);
-          indexByPopup[name] = features.length - 1;
-        }
-      } else if (PopupComponent) {
-        features.push(...featInfo.features);
-        featInfo.features.forEach(() => {
-          layers.push(featInfo.layer);
-          coordinates.push(featInfo.coordinate);
-        });
-      }
-    });
-    return { features, layers, coordinates };
-  }, [featureInfo]);
-
-  // When the featureIndex change we add the red circle then we pan on it.
-  useEffect(() => {
-    highlightLayer.getSource().clear();
-
-    // 'feature' can be a feature or an array
-    const feature = infoIndexed.features[featureIndex];
-    if (!feature) {
-      // When the user click on map to get new feature info, the infoIndexed is
-      // changed before the featureIndex. So the featureIndex is not reinitialized yet.
-      // It will be on the next render. So we just ignore if there is no feature to display.
-      return;
-    }
-    const features = Array.isArray(feature) ? feature : [feature];
-    const layerr = Array.isArray(infoIndexed.layers[featureIndex])
-      ? infoIndexed.layers[featureIndex][0]
-      : infoIndexed.layers[featureIndex];
-    const coordinate = infoIndexed.coordinates[featureIndex];
-    const coordinates = Array.isArray((coordinate || [])[0])
-      ? coordinate
-      : [coordinate];
-
-    features.forEach((feat, idx) => {
-      if (feat && feat.getGeometry()) {
-        const layerHighlightGeom =
-          layerr.get('getHighlightGeometry') &&
-          layerr.get('getHighlightGeometry')(feat, layerr, coordinates[idx]);
-        if (feat.getGeometry().getType() === GeometryType.POINT) {
-          highlightLayer
-            .getSource()
-            .addFeature(new Feature(layerHighlightGeom || feat.getGeometry()));
-        } else {
-          // In case mapbox render an icon for a polygon or a line we display
-          // the highlight style on the coordinate clicked.
-          // Needed for platforms layer.
-          const { layer: mbLayer } = feat.get('mapboxFeature') || {};
-          const defaultHighlightGeom =
-            mbLayer &&
-            mbLayer.layout &&
-            mbLayer.layout['icon-image'] &&
-            new Point(coordinates[idx]);
-
-          if (layerHighlightGeom || defaultHighlightGeom) {
-            highlightLayer
-              .getSource()
-              .addFeature(
-                new Feature(layerHighlightGeom || defaultHighlightGeom),
-              );
-          }
-        }
-      }
-    });
-
-    // We have to render the map otherwise the last selected features are displayed during animation.
-    map.renderSync();
-
-    // We want to recenter the map only if the coordinates clicked are under
-    // the Overlay (mobile and desktop) or Menu element (only desktop).
-    const coordinateClicked = coordinates[0];
-    if (!coordinateClicked) {
-      return;
-    }
-    panCenterFeature(map, [layerr], coordinateClicked, menuOpen, isMobile);
-  }, [map, isMobile, featureIndex, infoIndexed, menuOpen]);
+  // Hook to highlight map features
+  useHighlightLayer(featureInfo, highlightLayer, featureIndex);
 
   // The current feature(s) to display.
   const feature = infoIndexed.features[featureIndex];
