+++ src/menus/DirektverbindungenMenu/DvMenu.js
@@ -2,11 +2,15 @@
 import React, { useEffect, useMemo } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Divider, makeStyles } from '@material-ui/core';
+import VectorLayer from 'ol/layer/Vector';
+import VectorSource from 'ol/source/Vector';
 import DvFeatureInfo from '../../config/ch.sbb.direktverbindungen/DvFeatureInfo';
 import DvLayerSwitcher from './DvLayerSwitcher';
 import { setDisplayMenu } from '../../model/app/actions';
 import IframeMenu from '../IframeMenu';
 import useIsMobile from '../../utils/useIsMobile';
+import highlightPointStyle from '../../utils/highlightPointStyle';
+import useHighlightLayer from '../../utils/useHighlightLayer';
 
 const useStyles = makeStyles(() => {
   return {
@@ -15,6 +19,11 @@ const useStyles = makeStyles(() => {
     },
   };
 });
+
+const highlightLayer = new VectorLayer({
+  source: new VectorSource({ features: [] }),
+});
+highlightLayer.setStyle(highlightPointStyle);
 
 function DvMenu() {
   const dispatch = useDispatch();
@@ -57,6 +66,9 @@ function DvMenu() {
       zoomControls.style.display = isMobile ? 'none' : 'block';
     }
   }, [isMobile, topic, dispatch]);
+
+  // Hook to highlight map features
+  useHighlightLayer(featureInfo, highlightLayer);
 
   return (
     <IframeMenu
