+++ src/utils/getPopupComponent.js
@@ -1,0 +1,8 @@
+import popups from '../popups';
+
+const getPopupComponent = ({ popupComponent, layer }) => {
+  const comp = popupComponent || layer.get('popupComponent');
+  return typeof comp === 'string' ? popups[comp] : comp;
+};
+
+export default getPopupComponent;
